fix: 결제 API 기능 수정 #188

+++ first-server/src/main/java/com/hanhwa_tae/firstserver/payment/command/application/dto/request/CreateOrderHistoryRequest.java
@@ -8,7 +8,7 @@ import lombok.RequiredArgsConstructor;
 @Getter
 @RequiredArgsConstructor
 public class CreateOrderHistoryRequest {
-    private Long cartId;
+    private int cartId;
     private int price;
     private int quantity;
     private int targetId;
